Refactor project organization
