+++ site/src/components/course-planner/course-search/SectionListing.svelte
