+++ site/src/components/course-planner/schedule/ScheduleBackground.svelte
