+++ site/src/components/course-planner/schedule/ScheduleDay.svelte
