+++ site/src/components/course-planner/schedule/Tooltip.svelte
