+++ site/src/lib/CourseSearch.ts
